Move search files to src (#35822)
