+++ components/page-header/Breadcrumbs.tsx
@@ -22,7 +22,7 @@ export const Breadcrumbs = ({ inHeader }: Props) => {
       NOTE: The breadcrumbs class and the nav tag are used by the
       Lunr search scripts. The a tag generated by the Link is also used.
       If these change, please also change
-      updating script/search/parse-page-sections-into-records.js.
+      updating src/search/scripts/parse-page-sections-into-records.js.
     */
     <nav
       data-testid={inHeader ? 'breadcrumbs-header' : 'breadcrumbs-in-article'}
