+++ package.json
@@ -194,7 +194,7 @@
     "dev": "cross-env npm start",
     "fixture-dev": "cross-env ROOT=tests/fixtures npm start",
     "fixture-test": "cross-env ROOT=tests/fixtures npm test -- tests/rendering-fixtures",
-    "index-test-fixtures": "node script/search/index-elasticsearch.js -l en -l ja -V ghae -V dotcom --index-prefix tests -- tests/content/fixtures/search-indexes",
+    "index-test-fixtures": "node src/search/scripts/index-elasticsearch.js -l en -l ja -V ghae -V dotcom --index-prefix tests -- src/search/tests/fixtures/search-indexes",
     "lint": "eslint '**/*.{js,mjs,ts,tsx}'",
     "lint-translation": "cross-env NODE_OPTIONS=--experimental-vm-modules jest tests/linting/lint-files.js",
     "playwright-test": "playwright test --project=\"Google Chrome\"",
@@ -210,7 +210,7 @@
     "start-for-playwright": "cross-env ROOT=tests/fixtures NODE_ENV=test node server.js",
     "sync-search": "cross-env NODE_OPTIONS='--max_old_space_size=8192' start-server-and-test sync-search-server 4002 sync-search-indices",
     "sync-search-ghes-release": "cross-env GHES_RELEASE=1 start-server-and-test sync-search-server 4002 sync-search-indices",
-    "sync-search-indices": "script/search/sync-search-indices.js",
+    "sync-search-indices": "src/search/scripts/sync-search-indices.js",
     "sync-search-server": "cross-env NODE_ENV=production PORT=4002 MINIMAL_RENDER=true CHANGELOG_DISABLED=true node server.js",
     "translation-check": "start-server-and-test translation-check-server 4002 translation-check-test",
     "translation-check-server": "cross-env NODE_ENV=test PORT=4002 node server.js",
