+++ pages/[versionId]/search.tsx
@@ -1,4 +1,1 @@
-import Search from '../search'
-export { getServerSideProps } from '../search'
-
-export default Search
+export { default, getServerSideProps } from '../search'
