+++ src/search/components/NoQuery.tsx
