+++ src/search/components/SearchError.tsx
