+++ src/search/components/SearchResults.module.scss
