+++ src/search/components/index.tsx
@@ -7,11 +7,11 @@ import { sendEvent, EventType } from 'src/events/browser'
 import { useTranslation } from 'components/hooks/useTranslation'
 import { DEFAULT_VERSION, useVersion } from 'components/hooks/useVersion'
 import { useNumberFormatter } from 'components/hooks/useNumberFormatter'
-import type { SearchResultsT } from 'components/search/types'
-import { SearchResults } from 'components/search/SearchResults'
-import { SearchError } from 'components/search/SearchError'
-import { NoQuery } from 'components/search/NoQuery'
-import { Loading } from 'components/search/Loading'
+import type { SearchResultsT } from 'src/search/components/types'
+import { SearchResults } from 'src/search/components/SearchResults'
+import { SearchError } from 'src/search/components/SearchError'
+import { NoQuery } from 'src/search/components/NoQuery'
+import { Loading } from 'src/search/components/Loading'
 import { useQuery } from 'components/hooks/useQuery'
 import { usePage } from 'components/hooks/usePage'
 import { useMainContext } from 'components/context/MainContext'
