+++ src/search/components/types.ts
