+++ src/search/lib/config.js
