+++ src/search/scripts/parse-page-sections-into-records.js
