+++ src/search/scripts/popular-pages.js
