+++ src/search/scripts/sync.js
@@ -1,11 +1,11 @@
 #!/usr/bin/env node
 import chalk from 'chalk'
 
-import languages from '../../lib/languages.js'
+import languages from '../../../lib/languages.js'
 import buildRecords from './build-records.js'
 import findIndexablePages from './find-indexable-pages.js'
-import { allVersions } from '../../lib/all-versions.js'
-import { namePrefix } from '../../lib/search/config.js'
+import { allVersions } from '../../../lib/all-versions.js'
+import { namePrefix } from '../../../src/search/lib/config.js'
 import { writeIndexRecords } from './search-index-records.js'
 
 // Build a search data file for every combination of product version and language
