+++ src/search/tests/fixtures/search-indexes/github-docs-ghae-ja-records.json
