+++ src/search/tests/search.js
@@ -1,6 +1,6 @@
 import { expect, jest } from '@jest/globals'
 
-import { getDOM } from '../helpers/e2etest.js'
+import { getDOM } from '../../../tests/helpers/e2etest.js'
 
 describe('search results page', () => {
   jest.setTimeout(5 * 60 * 1000)
